Cleanup error paths, don't use hiface_chip_destroy()
The problem with hiface_chip_destroy() was this:

  - On disconnect ALSA is responsible of freeing the card by calling
    snd_card_free() itself, as requested by snd_card_free_when_closed(),
    which in turn calls hiface_dev_free().

  - hiface_dev_free() was then calling snd_card_free() again which
    resulted in recursion.

NOTE: in the error paths of hiface_chip_probe() we just need to call
snd_card_free() and not kfree(chip), as ALSA will take care of invoking
hiface_dev_free() for us which frees the chip data.

While at it, move the mutex_unlock(&register_mutex) right after
snd_card_register() as the mutex name suggests this protect multiple
registrations, and BTW snd-usb-audio does so too.

Signed-off-by: Antonio Ospite <ao2@amarulasolutions.com>