Fold hiface_chip_abort() into hiface_chip_disconnect() and remove it
hiface_chip_abort() was only called from hiface_chip_disconnect() and we
know that chip and card are defined when the latter is called, so the
pointer validity checks can be removed too.

No need to set chip->card to NULL as we are not going to use it again.

Signed-off-by: Antonio Ospite <ao2@amarulasolutions.com>