Simplify setting the pcm stream name
Other drivers like snd-usb-audio or snd-hda-intel do not set the pcm
stream name according to the actual hardware device name, but just
according to some properties of the stream itself (e.g. USB Audio or
VT1708B 8-Ch Digital).

We wan call the stream "USB-SPDIF Audio" in this case following what's
in the USB descriptor.

Now that the pcm stream name is not variable anymore, the signature of
hiface_pcm_init() can be simplified as well.

Signed-off-by: Antonio Ospite <ao2@amarulasolutions.com>