Make the code in hiface_pcm_pointer() more readable
Rename "ret" to "dma_offset" and use bytes_to_frames() so that who is
going to read the code has a semantic hint of what is going on here: the
function is returning the position in the DMA buffer, expressed in
frames.

That's exactly what the documentation says; on my first review and
before this change I didn't get what the calculation in the return
statement was for exactly.

Signed-off-by: Antonio Ospite <ao2@amarulasolutions.com>