Replace the swap_word() function with something more straightforward
I can see why the name swap_word() was chosen, words are indeed swapped
in this function, however having the signature showing a destination and
a source parameter it sounded like the data were swapped _between_ these
two buffers, and this is misleading.

The operation is more like a copy with swapped bytes, a better name
would be copy_swapped16() maybe, but it's even better to directly use
something resembling memcpy which is familiar to the reviewers.

The kernel offers a function to swap half of 32bit values which should
suit the case, its name is swahw32().

Signed-off-by: Antonio Ospite <ao2@amarulasolutions.com>